+++ pydocx/openxml/packaging/__init__.py
@@ -3,9 +3,9 @@ from pydocx.openxml.packaging.footnotes_part import FootnotesPart
 from pydocx.openxml.packaging.image_part import ImagePart
 from pydocx.openxml.packaging.main_document_part import MainDocumentPart
 from pydocx.openxml.packaging.numbering_definitions_part import NumberingDefinitionsPart  # noqa
-from pydocx.openxml.packaging.OpenXmlPackage import OpenXmlPackage
-from pydocx.openxml.packaging.OpenXmlPart import OpenXmlPart
-from pydocx.openxml.packaging.OpenXmlPartContainer import OpenXmlPartContainer
+from pydocx.openxml.packaging.open_xml_package import OpenXmlPackage
+from pydocx.openxml.packaging.open_xml_part import OpenXmlPart
+from pydocx.openxml.packaging.open_xml_part_container import OpenXmlPartContainer  # noqa
 from pydocx.openxml.packaging.style_definitions_part import StyleDefinitionsPart  # noqa
 from pydocx.openxml.packaging.word_processing_document import WordprocessingDocument  # noqa
