+++ pydocx/openxml/packaging/main_document_part.py
@@ -9,7 +9,7 @@ from pydocx.openxml.packaging.font_table_part import FontTablePart
 from pydocx.openxml.packaging.footnotes_part import FootnotesPart
 from pydocx.openxml.packaging.image_part import ImagePart
 from pydocx.openxml.packaging.numbering_definitions_part import NumberingDefinitionsPart  # noqa
-from pydocx.openxml.packaging.OpenXmlPart import OpenXmlPart
+from pydocx.openxml.packaging.open_xml_part import OpenXmlPart
 from pydocx.openxml.packaging.style_definitions_part import StyleDefinitionsPart  # noqa
